Split up encoder and decoder for simplicity.
